requirements.txt
Flask==3.0.3
sentence-transformers==2.7.0
faiss-cpu==1.8.0
PyPDF2==3.0.1
python-docx==1.1.2
requests==2.31.0
werkzeug==3.0.3
